phone_detect/requirements.txt
# Core Dependencies
torch>=1.12.0
torchvision>=0.13.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0

# Model Libraries
timm>=0.6.0
ultralytics>=8.0.0
segmentation-models-pytorch>=0.3.3

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.64.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Data Handling
pandas>=1.4.0

# Optional: For better performance
# tensorboard>=2.10.0  # Uncomment if you want to use TensorBoard

